requirements.txt
numpy
stt-tflite
pyaudio
webrtcvad
halo
